refactor(protoncalendar): extract travel-duration helpers to shared icsCalendarUtils
getTravelDurationInSeconds and applyTravelDuration are generic ICS/Apple
Travel Time utilities with no Proton-specific logic. Move them to the
shared _utils/calendars/icsCalendarUtils.ts and update both
protoncalendar and ics-feedcalendar to import from there.
